feat: Show Main page
created the ShowMain page and moved the logic from ShowMainInfo
